Added vulkan RAII helpers, Instance, InstanceBuilder

+++ src/main.cpp
@@ -33,7 +33,7 @@ core::Address readAddress() {
         return core::Address::localhost(20'040);
     }
     std::string const ip = line.substr(0, delim);
-    uint16_t const port = std::stoul(line.substr(delim + 1));
+    uint16_t const port = static_cast<uint16_t>(std::stoul(line.substr(delim + 1)));
     return core::Address::make(ip, port);
 }
